Implement classes for recipe search story
This commit is a big large, but it revolves around implementing everything needed in order for a user to be able to search recipes within the database.

- Recipe class elements
- Ingredient class elements
- MultimediaContent class elements
- Data seeding elements
- Application and MainApplicationController elements
- Basic view templates, a style sheet, and simple javascript

Right now, seeding occurs on a new load of the application. This means that it will seed this data to the remote database everytime the application is run. Will need to look into this.
Additionally, the recipe steps is not behaving correctly because it is storing the hash of the list of strings, not the content of the strings themselves.
Still much to do!